aligned the home page & deleted the waste file and folder
