+++ package.json
@@ -15,8 +15,9 @@
     "@testing-library/user-event": "^12.8.3",
     "animate.css": "^4.1.1",
     "bootstrap": "^5.3.6",
-    "framer-motion": "^12.4.10",
+    "framer-motion": "^11.0.17",
     "lucide-react": "^0.511.0",
+    "motion-dom": "^12.16.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.10",
     "react-countup": "^6.1.1",
