+++ src/components/About_us/About_us.jsx
@@ -215,12 +215,6 @@ function App() {
                         </div>
                     </div>
                 </section>
-
-                {/* Footer Section (commented out as in your original code) */}
-                {/* <footer className="bg-gray-800 text-white py-8 text-center px-4 md:px-8">
-                    <p>&copy; {currentYear} Intellimindz. All rights reserved.</p>
-                    <p className="text-sm text-gray-400 mt-1">Website Clone for Demonstration</p>
-                </footer> */}
             </div>
         </>
     );
