--- src/components/Contactus/Intro.css
@@ -1,174 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=DM+Sans:ital,wght@0,100..1000;1,100..1000&display=swap');
-
-body {
-  font-family: 'DM Sans', sans-serif;
-}
-
-.contact-hero {
-  background: url('../../assests/images/contactintro.png') no-repeat center center/cover;
-  color: white;
-  padding: 200px 20px;
-  text-align: center;
-  position: relative;
-}
-
-.overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 0%;
-  background: rgba(0, 0, 0, 0.4);
-  z-index: 1;
-}
-
-.contact-hero h1 {
-  position: relative;
-  z-index: 2;
-}
-
-.contact-hero h1 span {
-  font-weight: 700;
-}
-
-.contact-main {
-  display: flex;
-  flex-wrap: wrap;
-  padding: 40px 10%;
-  background: #f5f5f5;
-}
-
-.contact-left,
-.contact-right {
-  flex: 1 1 45%;
-  margin: 20px;
-}
-
-.contact-left h2,
-.contact-right h2 {
-  font-size: 24px;
-  margin-bottom: 10px;
-}
-
-.contact-left p {
-  font-size: 16px;
-  margin-bottom: 10px;
-  color: #333;
-}
-
-.contact-left .highlight {
-  color: #007bff;
-  font-weight: bold;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-}
-
-form input,
-form select,
-form textarea {
-  margin-bottom: 15px;
-  padding: 10px;
-  font-size: 15px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.mobile-group {
-  display: flex;
-  gap: 10px;
-}
-
-form button {
-  background-color: #007bff;
-  color: white;
-  font-weight: bold;
-  padding: 12px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-form button:hover {
-  background-color: #0056b3;
-}
-
-.location-intro {
-  text-align: center;
-  padding: 40px 20px;
-  font-family: 'DM Sans', sans-serif;
-  background-color: #fff;
-}
-
-.location-text h2 {
-  color: #0c4c65;
-  font-size: 2rem;
-  font-weight: bold;
-  margin-bottom: 10px;
-}
-
-.location-text h4 {
-  font-weight: bold;
-  font-size: 16px;
-  margin-bottom: 20px;
-}
-
-.location-text p {
-  color: #333;
-  max-width: 800px;
-  margin: 0 auto 40px auto;
-  line-height: 1.6;
-}
-
-.contact-location-card {
-  background-color: #d3d3d3;
-  box-shadow: 0px 2px 20px rgba(0, 0, 0, 0.1);
-  padding: 25px 20px;
-  max-width: 500px;
-  margin: 0 auto;
-  border-radius: 12px;
-  display: flex;
-  align-items: flex-start;
-  gap: 15px;
-}
-
-.location-icon {
-  font-size: 28px;
-  color: white;
-  background-color: #0c4c65;
-  border-radius: 50%;
-  padding: 12px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 50px;
-  width: 50px;
-}
-
-.contact-location-card h3 {
-  margin: 0;
-  font-size: 20px;
-  font-weight: bold;
-  color: #0c4c65;
-}
-
-.contact-location-card p {
-  margin: 5px 0 0;
-  font-size: 15px;
-  color: #333;
-}
-
-.location-card-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 30px;
-  justify-content: center;
-  max-width: 900px;
-  margin: 0 auto;
-}
-
-.contact-location-card {
-  margin: 0;
-}
