--- src/components/Feature-section/features.css
@@ -1,7 +1,0 @@
-.single__feature h2 i {
-  font-size: 2.5rem;
-}
-
-.single__feature h6 {
-  font-weight: 600;
-}
