--- src/components/Free-course-section/FreeCourse.css
@@ -1,35 +1,0 @@
-.training-approach {
-  text-align: center;
-  padding: 20px;
-  background-color: #f9f9f9;
-  border-radius: 8px;
-  margin: 20px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.training-approach h2 {
-  color: #333;
-  font-size: 24px;
-  margin-bottom: 20px;
-}
-
-.training-approach ul {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.training-approach ul li {
-  background-color: #fff;
-  margin: 10px 0;
-  padding: 15px;
-  border-radius: 4px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  font-size: 18px;
-  color: #555;
-}
-
-.training-approach ul li:hover {
-  background-color: #f1f1f1;
-  cursor: pointer;
-}
