--- src/components/Hero-Section/message-icon.css
@@ -1,25 +1,0 @@
-.message-icon-container {
-    position: fixed;
-    left: 20px;
-    bottom: 50px;
-    z-index: 1000;
-  }
-  
-  .message-icon {
-    background-color: rgb(24, 253, 3);
-    padding: 10px;
-    border-radius: 50%;
-    box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2);
-    cursor: pointer;
-    transition: transform 0.3s ease-in-out;
-  }
-  
-  .message-icon:hover {
-    transform: scale(1.1);
-  }
-  
-  .message-icon-img {
-    width: 40px;
-    height: 40px;
-  }
-
