+++ src/components/Home/TrainingOptions.css
@@ -34,7 +34,7 @@
     padding: -10px;
     width: 270px;
     text-align: center;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.3);
     transition: transform 0.3s ease-in-out;
 }
 
@@ -55,7 +55,7 @@
 .training-header {
     font-size: 1.4rem;
     font-weight: bold;
-    color: #b49b49;
+    color: #d1a617;
     margin-bottom: 20px;
 }
