+++ src/components/Home/choose-us.css
@@ -1,12 +1,24 @@
 .choose-box {
   background-color: #ffdab9;
-  /* Light orange background */
   padding: 20px;
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(8, 0, 0, 0.1);
   text-align: center;
   margin-top: 20px;
-  /* Move the box higher */
+  gap: 10px;
+}
+
+.choose-section-title{
+    font-size: 30px;
+  font-weight: bold;
+  color: #0a0a0a;
+  margin-bottom: 15px;
+  text-align: center;
+}
+
+.feature-description{
+  color: #0e0e0e !important;
+  font-size: 15px !important;
 }
 
 .text-center mb-4 {
@@ -15,20 +27,16 @@
 
 .choose-box h4 {
   color: #ff4500;
-  /* Orange color for the header */
 }
 
 .choose-box p {
   color: #000000;
-  /* Black color for the description */
 }
 
-/* Apply font to the entire choose-us section */
 .choose-us {
   font-family: "DM Sans", sans-serif;
 }
 
-/* Apply font to specific text elements */
 .choose-us h2,
 .choose-us h4,
 .choose-us p {
@@ -78,20 +86,14 @@
   padding: 10px;
 }
 
-/* Enlarge Icons */
 .feature-icon {
   font-size: 3rem;
-  /* Increase the size of the icons */
   color: #000000;
-  /* Optional: Change the color of the icons */
   margin-bottom: 15px;
-  /* Add some spacing below the icons */
 }
 
-/* Existing Styles */
 .choose-box {
   background-color: #ffdab9;
-  /* Light orange background */
   padding: 20px;
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
@@ -100,12 +102,10 @@
 
 .choose-box h4 {
   color: #ff4500;
-  /* Orange color for the header */
 }
 
 .choose-box p {
   color: #000000;
-  /* Black color for the description */
 }
 
 .choose__img img {
@@ -151,19 +151,12 @@
   padding: 10px;
 }
 
-
-
-/* Enlarge Icons */
 .feature-icon {
   font-size: 3rem;
-  /* Increase the size of the icons */
   color: #000000;
-  /* Optional: Change the color of the icons */
   margin-bottom: 20px;
-  /* Add some spacing below the icons */
 }
 
-/* Transform Effect for Boxes */
 .choose-box {
   text-align: center;
   padding: 20px;
@@ -171,26 +164,21 @@
   border-radius: 10px;
   box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
   transition: transform 0.3s ease-in-out;
-  /* Add smooth transition */
 }
 
 .choose-box h4 {
   color: #ff4500;
-  /* Orange color for the header */
 }
 
 .choose-box p {
   color: #000000;
-  /* Black color for the description */
   font-size: 1.2rem;
 }
 
 .choose-box:hover {
   transform: translateY(-30px);
-  /* Move the box up on hover */
 }
 
-/* Existing Styles */
 .choose__img img {
   border-radius: 15px;
 }
@@ -234,24 +222,12 @@
   padding: 10px;
 }
 
-
-
-/* homeimage - Updated for alignment */
 .img-fluid {
-  max-width: 550px !important;
+  max-width: 490px !important;
   margin-left: 100px !important;
-  /* padding-left: -150px !important; */
-  /* Ensures responsiveness */
   background-color: #17bf9e;
-    /* box-shadow: 0 2px 4px rgba(8, 0, 0, 0.1); */
-  height: 500px !important;
-  /* Maintains aspect ratio */
+  height: 450px !important;
   display: block;
-  /* Removes extra space below the image */
   margin: 0 auto;
-  /* Centers the image horizontally within its column */
   border-radius: 15px;
-  /* Apply border-radius for consistency with other images */
-  /* Remove specific height and width to let it scale naturally */
-  /* margin-top and margin-left will be managed by the parent column's flexbox alignment */
 }
