+++ src/components/Home/testimonial.css
@@ -10,7 +10,7 @@
 .section-title {
   font-size: 30px;
   font-weight: bold;
-  color: #070707;
+  color: #111111;
   margin-bottom: 15px;
   margin-left: 400px;
 }
