+++ src/components/SAP/Header/Introsap.css
@@ -7,7 +7,7 @@
 
 .intro-container {
   width: 100%;
-  height: 70vh; /* Full viewport height */
+  height: 100vh; /* Full viewport height */
   display: flex;
   align-items: center;
   justify-content: center;
@@ -27,12 +27,15 @@
 .intro-content {
   position: absolute;
   z-index: 2;
-  margin-top: 40px;
+  /* text-align: center; */
+  margin-top: 200px;
+  margin-left: 100px;
 }
 
 .intro-content h1 {
   font-size: 2rem;
   font-weight: 400;
+    text-align: center;
   margin-right: 600px;
 }
